Move DescopeFlow into its own file
